started safe add test

+++ test-data/PrimalityCheck.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.9;
 
 contract PrimalityCheck {
-    function factor(uint x, uint y) public pure returns(uint) {
-      require(1 < x && x < 973013 && 1 < y && y < 973013);
-      assert(x*y != 973013);
-      return 1337;
+    function factor(uint256 x, uint256 y) public pure returns (uint256) {
+        require(1 < x && x < 973013 && 1 < y && y < 973013);
+        assert(x * y != 973013);
+        return 1337;
     }
 }
